+++ src/VoiceInput.Native/native_entry.cpp
@@ -1,6 +1,0 @@
-#include <windows.h>
-
-extern "C" __declspec(dllexport) int VoiceInputNativeVersion()
-{
-    return 1;
-}